refactor: renamed directory to model
Relates #166
