+++ components/form/ProfileEdit.tsx
@@ -1,5 +1,5 @@
 'use client';
-import editRow from '@/supabase/modals/editRow';
+import editRow from '@/supabase/models/editRow';
 import { PartialItem, editProfile } from '@/types/types';
 import { useEffect, useState } from 'react';
 import { useForm } from 'react-hook-form';
