+++ components/messaging/CurrentConversation.tsx
@@ -6,22 +6,20 @@ import MessageForm from './MessageForm';
 import { ConversationPartner } from './ConversationPartner';
 import { useEffect, useState, useRef } from 'react';
 import { useConversationContext } from '../../context/conversationContext';
-import selectMessagesByConversationId from '@/supabase/modals/messaging/selectMessagesByConversationId';
+import selectMessagesByConversationId from '@/supabase/models/messaging/selectMessagesByConversationId';
 import {
   formatTimeMarker,
   formatDateMarker,
 } from '../../utils/formatTimeStamp';
+import newClient from '@/supabase/utils/newClient';
 
 const CurrentConversation: React.FC = () => {
-  const {
-    allConversations,
-    currentConversation,
-    setCurrentConversation,
-    supabase,
-  } = useConversationContext();
+  const { allConversations, currentConversation, setCurrentConversation } =
+    useConversationContext();
   const [currentMessages, setCurrentMessages] = useState<MessageType[]>([]);
   const [isScrolling, setIsScrolling] = useState<boolean>(false);
   const chatWindowRef = useRef<HTMLDivElement>(null);
+  const supabase = newClient();
 
   useEffect(() => {
     setCurrentConversation && setCurrentConversation(allConversations[0]);
@@ -63,7 +61,7 @@ const CurrentConversation: React.FC = () => {
     return () => {
       supabase.removeChannel(channel);
     };
-  }, [supabase, currentMessages, setCurrentMessages]);
+  }, [currentMessages, setCurrentMessages]);
 
   useEffect(() => {
     const handleScroll = () => {
