+++ context/conversationContext.tsx
@@ -1,5 +1,4 @@
 'use client';
-import { SupabaseClient } from '@supabase/supabase-js';
 import {
   Dispatch,
   SetStateAction,
@@ -11,7 +10,6 @@ import {
   AllConversationsType,
   ConversationCardType,
 } from '@/types/messagingTypes';
-import newClient from '@/supabase/utils/newClient';
 
 type ConverstaionContextProviderProps = {
   children: React.ReactNode;
@@ -28,7 +26,6 @@ type ConversationContext = {
   setShowConversationsList: Dispatch<SetStateAction<boolean>>;
   currentUserId: string;
   setCurrentUserId: Dispatch<SetStateAction<string>>;
-  supabase: SupabaseClient;
 };
 
 export const ConversationContext = createContext<ConversationContext | null>(
@@ -49,8 +46,6 @@ export default function ConversationContextProvider({
 
   const [currentUserId, setCurrentUserId] = useState<string>('');
 
-  const supabase = newClient();
-
   return (
     <ConversationContext.Provider
       value={{
@@ -62,7 +57,6 @@ export default function ConversationContextProvider({
         setShowConversationsList,
         currentUserId,
         setCurrentUserId,
-        supabase,
       }}
     >
       {children}
