+++ supabase/models/deleteItems.ts
