+++ supabase/models/searchByName.ts
